move location file
