+++ app/src/main/java/de/tum/in/tumcampusapp/component/ui/cafeteria/CafeteriaLocationDao.java
@@ -7,7 +7,7 @@ import androidx.room.Query;
 
 import java.util.List;
 
-import de.tum.in.tumcampusapp.component.ui.cafeteria.model.Location;
+import de.tum.in.tumcampusapp.component.ui.openinghour.model.Location;
 
 @Dao
 public interface CafeteriaLocationDao {
